GRAD2-1931 Distribution PSI FTP Part 12 work. Introduced separate methods for folder paths and file paths for PSIRUNs
