+++ api/src/main/java/ca/bc/gov/educ/api/distribution/util/IOUtils.java
@@ -67,62 +67,4 @@ public class IOUtils {
         }
     }
 
-    /**
-     * Grad2-1931 : Creates folder structure in temp directory for files to be placed basing on the DISTRUN and PSIRUN - mchintha
-     *
-     * @param processorData is ProcessorData object
-     * @param minCode is PSI code
-     * @param schoolCategoryCode is schoolCategoryCode
-     * @return StringBuilder
-     */
-    //
-    public static StringBuilder createFolderStructureInTempDirectory(ProcessorData processorData, String minCode, String schoolCategoryCode) {
-        String districtCode = StringUtils.substring(minCode, 0, 3);
-        String activityCode = processorData.getActivityCode();
-        String transmissionMode = processorData.getTransmissionMode();
-        StringBuilder directoryPathBuilder = new StringBuilder();
-        StringBuilder filePathBuilder = new StringBuilder();
-        Path path;
-        try {
-            Boolean conditionResult = (EducDistributionApiConstants.MONTHLYDIST.equalsIgnoreCase(activityCode) || "02".equalsIgnoreCase(schoolCategoryCode));
-            if(!transmissionMode.isBlank() && (transmissionMode.equalsIgnoreCase(EducDistributionApiConstants.TRANSMISSION_MODE_FTP) || (transmissionMode.equalsIgnoreCase(EducDistributionApiConstants.TRANSMISSION_MODE_PAPER)))) {
-
-                if (Boolean.TRUE.equals(conditionResult)) {
-                    directoryPathBuilder.append(EducDistributionApiConstants.TMP_DIR).append(EducDistributionApiConstants.FILES_FOLDER_STRUCTURE).append(transmissionMode.toUpperCase()).append(EducDistributionApiConstants.DEL).append(processorData.getBatchId()).append(EducDistributionApiConstants.DEL).append(minCode).append(EducDistributionApiConstants.DEL);
-
-                } else {
-                    directoryPathBuilder.append(EducDistributionApiConstants.TMP_DIR).append(EducDistributionApiConstants.FILES_FOLDER_STRUCTURE).append(transmissionMode.toUpperCase()).append(EducDistributionApiConstants.DEL).append(processorData.getBatchId()).append(EducDistributionApiConstants.DEL).append(districtCode).append(EducDistributionApiConstants.DEL).append(minCode);
-                }
-                path = Paths.get(directoryPathBuilder.toString());
-                Files.createDirectories(path);
-
-                if (Boolean.TRUE.equals(conditionResult)) {
-                    filePathBuilder.append(EducDistributionApiConstants.TMP_DIR).append(EducDistributionApiConstants.FILES_FOLDER_STRUCTURE).append(transmissionMode.toUpperCase()).append(EducDistributionApiConstants.DEL).append(processorData.getBatchId()).append(EducDistributionApiConstants.DEL).append(minCode).append(EducDistributionApiConstants.DEL);
-                } else {
-                    filePathBuilder.append(EducDistributionApiConstants.TMP_DIR).append(EducDistributionApiConstants.FILES_FOLDER_STRUCTURE).append(transmissionMode.toUpperCase()).append(EducDistributionApiConstants.DEL).append(processorData.getBatchId()).append(EducDistributionApiConstants.DEL).append(districtCode).append(EducDistributionApiConstants.DEL).append(minCode);
-                }
-            }
-            else {
-                if (Boolean.TRUE.equals(conditionResult)) {
-                    directoryPathBuilder.append(EducDistributionApiConstants.TMP_DIR).append(processorData.getBatchId()).append(EducDistributionApiConstants.DEL).append(minCode).append(EducDistributionApiConstants.DEL);
-                } else {
-                    directoryPathBuilder.append(EducDistributionApiConstants.TMP_DIR).append(processorData.getBatchId()).append(EducDistributionApiConstants.DEL).append(districtCode).append(EducDistributionApiConstants.DEL).append(minCode);
-                }
-                path = Paths.get(directoryPathBuilder.toString());
-                Files.createDirectories(path);
-
-                if (Boolean.TRUE.equals(conditionResult)) {
-                    filePathBuilder.append(EducDistributionApiConstants.TMP_DIR).append(processorData.getBatchId()).append(EducDistributionApiConstants.DEL).append(minCode);
-                } else {
-                    filePathBuilder.append(EducDistributionApiConstants.TMP_DIR).append(processorData.getBatchId()).append(EducDistributionApiConstants.DEL).append(districtCode).append(EducDistributionApiConstants.DEL).append(minCode);
-                }
-
-            }
-        } catch (Exception e) {
-            logger.error(EXCEPTION, e.getLocalizedMessage());
-        }
-
-        return filePathBuilder;
-    }
-
-}
+ }
